+++ geniuspace/database/seeders/DatabaseSeeder.php
@@ -13,5 +13,6 @@ class DatabaseSeeder extends Seeder
         $this->call(PlatformSeeder::class);
         $this->call(GodSeeder::class);
         $this->call(MoatSeeder::class);
+        $this->call(PilotSeeder::class);
     }
 }